+++ source/commands/club-skip-meeting.js
@@ -2,7 +2,7 @@ const { CommandWrapper } = require('../classes/index.js');
 const { cancelClubEvent, createClubEvent, clearClubReminder, scheduleClubReminderAndEvent, updateClubDetails } = require('../engines/clubEngine.js');
 const { isClubHostOrModerator } = require('../engines/permissionEngine.js');
 const { getClubDictionary, updateClub, updateList } = require('../engines/referenceEngine.js');
-const { timeConversion } = require('../helpers.js');
+const { timeConversion } = require('../util/mathUtil.js');
 
 const mainId = "club-skip-meeting";
 module.exports = new CommandWrapper(mainId, "Skip the next club meeting, cancelling/resetting reminders", null, false, 3000,
