+++ source/util/mathUtil.js
@@ -1,0 +1,55 @@
+/** Convert an amount of time from a starting unit to a different one
+ * @param {number} value
+ * @param {"w" | "d" | "h" | "m" | "s" | "ms"} startingUnit
+ * @param {"w" | "d" | "h" | "m" | "s" | "ms"} resultUnit
+ * @returns {number}
+ */
+function timeConversion(value, startingUnit, resultUnit) {
+	const unknownUnits = [];
+	let msPerStartUnit = 1;
+	switch (startingUnit.toLowerCase()) {
+		case "w":
+			msPerStartUnit *= 7;
+		case "d":
+			msPerStartUnit *= 24;
+		case "h":
+			msPerStartUnit *= 60;
+		case "m":
+			msPerStartUnit *= 60;
+		case "s":
+			msPerStartUnit *= 1000;
+		case "ms":
+			msPerStartUnit *= 1;
+			break;
+		default:
+			unknownUnits.push(startingUnit);
+	}
+
+	let msPerResultUnit = 1;
+	switch (resultUnit.toLowerCase()) {
+		case "w":
+			msPerResultUnit *= 7;
+		case "d":
+			msPerResultUnit *= 24;
+		case "h":
+			msPerResultUnit *= 60;
+		case "m":
+			msPerResultUnit *= 60;
+		case "s":
+			msPerResultUnit *= 1000;
+		case "ms":
+			msPerResultUnit *= 1;
+			break;
+		default:
+			unknownUnits.push(resultUnit);
+	}
+	if (!unknownUnits.length) {
+		return value * msPerStartUnit / msPerResultUnit;
+	} else {
+		throw new Error(`Unknown unit used: ${unknownUnits.join(", ")} (allowed units: ms, s, m, h, d, w)`)
+	}
+}
+
+module.exports = {
+	timeConversion
+}
